+++ src/contents/LottoConstants.js
